desafio/requirements.txt
asgiref==3.3.1
Django==3.1.4
djangorestframework==3.12.2
drf-generators==0.5.0
pytz==2020.4
sqlparse==0.4.1
